build unminified socket distribution files
