+++ package.json
@@ -51,6 +51,7 @@
     "socket.io": "2.0.4",
     "strip-loader": "0.1.2",
     "text-blob-builder": "0.0.1",
+    "webpack-merge": "4.1.2",
     "webpack-stream": "3.2.0",
     "zuul": "^3.11.1  ",
     "zuul-builder-webpack": "^1.2.0",
