+++ support/webpack.config.dev.js
@@ -1,0 +1,39 @@
+
+var webpack = require('webpack');
+
+module.exports = {
+  name: 'default',
+  entry: './lib/index.js',
+  output: {
+    library: 'io',
+    libraryTarget: 'umd',
+    filename: 'socket.io.dev.js'
+  },
+  externals: {
+    global: glob()
+  },
+  devtool: 'source-map',
+  module: {
+    loaders: [{
+      test: /\.js$/,
+      exclude: /(node_modules|bower_components)/,
+      loader: 'babel', // 'babel-loader' is also a legal name to reference
+      query: { presets: ['es2015'] }
+    }, {
+      test: /\json3.js/,
+      loader: 'imports?define=>false'
+    }]
+  }
+};
+
+/**
+ * Populates `global`.
+ *
+ * @api private
+ */
+
+function glob () {
+  return 'typeof self !== "undefined" ? self : ' +
+    'typeof window !== "undefined" ? window : ' +
+    'typeof global !== "undefined" ? global : {}';
+}
