+++ support/webpack.config.js
@@ -1,18 +1,13 @@
-
 var webpack = require('webpack');
+var merge = require('webpack-merge');
+var baseConfig = require('./webpack.config.dev.js')
 
-module.exports = {
-  name: 'default',
-  entry: './lib/index.js',
+module.exports = merge(baseConfig, {
   output: {
     library: 'io',
     libraryTarget: 'umd',
     filename: 'socket.io.js'
   },
-  externals: {
-    global: glob()
-  },
-  devtool: 'source-map',
   plugins: [
     new webpack.optimize.UglifyJsPlugin({
       compress: {
@@ -27,27 +22,4 @@ module.exports = {
       }
     })
   ],
-  module: {
-    loaders: [{
-      test: /\.js$/,
-      exclude: /(node_modules|bower_components)/,
-      loader: 'babel', // 'babel-loader' is also a legal name to reference
-      query: { presets: ['es2015'] }
-    }, {
-      test: /\json3.js/,
-      loader: 'imports?define=>false'
-    }]
-  }
-};
-
-/**
- * Populates `global`.
- *
- * @api private
- */
-
-function glob () {
-  return 'typeof self !== "undefined" ? self : ' +
-    'typeof window !== "undefined" ? window : ' +
-    'typeof global !== "undefined" ? global : {}';
-}
+});
