+++ support/webpack.config.slim.js
@@ -1,47 +1,14 @@
-
 var webpack = require('webpack');
+var merge = require('webpack-merge');
+var baseConfig = require('./webpack.config.slim.dev.js')
 
-module.exports = {
-  name: 'slim',
-  entry: './lib/index.js',
+module.exports = merge(baseConfig, {
   output: {
     library: 'io',
     libraryTarget: 'umd',
     filename: 'socket.io.slim.js'
   },
-  externals: {
-    global: glob(),
-    json3: 'JSON'
-  },
-  devtool: 'source-map',
   plugins: [
-    new webpack.NormalModuleReplacementPlugin(/debug/, process.cwd() + '/support/noop.js'),
     new webpack.optimize.UglifyJsPlugin()
   ],
-  module: {
-    loaders: [{
-      test: /\.js$/,
-      exclude: /(node_modules|bower_components)/,
-      loader: 'babel', // 'babel-loader' is also a legal name to reference
-      query: { presets: ['es2015'] }
-    }, {
-      test: /\json3.js/,
-      loader: 'imports?define=>false'
-    }, {
-      test: /\.js$/,
-      loader: 'strip-loader?strip[]=debug'
-    }]
-  }
-};
-
-/**
- * Populates `global`.
- *
- * @api private
- */
-
-function glob () {
-  return 'typeof self !== "undefined" ? self : ' +
-    'typeof window !== "undefined" ? window : ' +
-    'typeof global !== "undefined" ? global : {}';
-}
+});
